fix: 점포 대시보드 버그 수정 #49

+++ src/main/java/org/example/luckyburger/domain/review/service/ReviewEntityFinder.java
@@ -31,7 +31,7 @@ public class ReviewEntityFinder {
         return reviewRepository.findAllByShop(shop);
     }
 
-    public Integer getSumOfRatingByShop(Shop shop) {
+    public Double getSumOfRatingByShop(Shop shop) {
         return reviewRepository.findSumOfRatingByShop(shop);
     }
